pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadep"
version = "0.1.0"
description = "Cross-Domain Adversarial Expert Panel — grounded adversarial review for high-stakes decisions"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1",
    "pydantic>=2.0",
    "scikit-learn>=1.4",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
]

[project.scripts]
cadep = "cadep.cli:cli"

[tool.setuptools.packages.find]
include = ["cadep*"]
